FRAS/requirements.txt
# minimal dependencies for Face Recognition Attendance System
# remove outdated / unnecessary packages; add extras as needed
opencv-contrib-python>=4.1
numpy>=1.17
pandas>=1.0
pillow>=8.0
yagmail>=0.15
# optional packages:
# demjson3 for legacy JSON handling
# demjson3>=3.0
# add other libs if your code imports them
